move callFunctionOnJSModule to public api (#37399)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37399

Changelog: [Internal]

in this stack, i remove the `ReactInstanceForwarding` protocol. it is super lean and is only used by two classes, and the polymorphic behavior never comes into play because we never have a pointer to `id<ReactInstanceForwarding>` in our codebase.

being able to call into JS from native is tablestakes behavior in userland, so i'm moving that to the public API. i also renamed it to be more clear that it's calling from native into JS, not the other way around.

Reviewed By: cipolleschi

Differential Revision: D45760090

fbshipit-source-id: 8ac99723796cb65891076e8e7d69aeefe3e94213