introduce callback path for js errors (#37343)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37343

Changelog: [Internal]

in this stack i address the `jsErrorHandlingFunc` argument in the constructor of `RCTHost`.

currently in userland, the delegate of `RCTHost` is responsible for the following if they want to handle the errors fired in C++ land:
- importing the C++ class MapBuffer
- creating a C++ lambda method
- parsing the MapBuffer

in this diff stack i simplify this so we only need to implement a delegate method where userland only receives Foundation types. yay!

bypass-github-export-checks

Reviewed By: RSNara

Differential Revision: D45720131

fbshipit-source-id: b5541abe505140b4b85cbec1a8acecc1a3df7a72