Avoid usage of JSI Specializations which may use RTTI
Summary:
`getNativeState()` and the version of `getHostObject()` not specialized to `HostObject` do a dynamic_cast inside of an assert() to do a type check.

Most of RN for Android is built without RTTI enabled, but hadn't run into issues with these because fbandroid sets NDEBUG even in debug builds, causing any raw assert() to be compiled out.

All these functions really do is a checked cast, where the check is only in debug, and already compiled out on Android. This changes them to do a static_cast themselves instead.

This solution is probably preferable to conditioning the existing asserts on RTTI availability within JSI, since some libraries do build with RTTI, and changing behavior dependent on it in a shared header seems like a recipe for ODR. Though I'm not sure why casting is the API responsibility of JSI in general.

Static asserts are added so that new usages of functions meant to use RTTI are not added to units compiled internally with `-frtti`.

Changelog: [Internal]

Differential Revision: D45814500

fbshipit-source-id: 99d8dedd0350af4d625969bc7a422ae31d882463