Add support to register New Renderer components from the AppDelegate. (#37273)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37273

When using dynamic frameworks, we can't rely on Codegen to register all the components into the renderer. That's because, we would have to codegen a new target, which depends on ALL the 3rd party dependencies that expose a UI component.

The previous PR adds support for distributed automatic registration when components are loaded in memory.

However, not to slow down the adoption of the New Architecture, there could be apps that need to register a component that does not support the distributed approach yet. Thanks to this method, apps can register those components.

## Changelog:
[iOS][Added] - Added a mechanism to register components into the renderer from the user app.

Differential Revision: https://www.internalfb.com/diff/D45605688?entry_point=27

fbshipit-source-id: b68e8582a6c220d79236ab9425d46f45c6b0f3f1