Ensure not to exceed the Spanned priority limit
...as it led to totally random results for deep text node trees.

Instead, assign high priority values to SetSpanOperations with the
highest priority (which represent the nodes high in the tree), but clamp
the priority 0 for SetSpanOperations with the lowest priority (which
represent nodes closer to the leaves).

This isn't everything that could be done in this area, but this is a
relatively non-invasive change which fixes an obvious bitwise overflow
bug.