decouple RCTHost bundleURL setter from delegate (#37566)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37566

Changelog: [Internal]

in this diff stack, i remove `- (NSURL *)getBundleURL;` from the bridgeless API.

the reason for this is because the bundle url does not change in the middle of an app session, unless for developer mode, so it's more robust for us to pass it as a dependency.

at the end of the diff stack, i'll add an escape hatch that allows us to update the bundle url for our dev mode case, so we can delete the API.

in this diff, i do some cleanup so the bundle logic is more reusable.

Reviewed By: sammy-SC

Differential Revision: D45937853

fbshipit-source-id: bffe808a26e4cff0de558f6f8fc96f45d80c934d