introduce a private API for RCTHost to refresh its bundle url (#37567)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37567

Changelog: [Internal]

in the situation where we want to refresh the bundle url due to developer mode, we add a private API that allows the host to refresh its bundle URL if needed. this private API will be called on cmd + R, which is the only case that supports the changing bundle URL in the middle of the app session.

Reviewed By: cipolleschi

Differential Revision: D45937856

fbshipit-source-id: 718047abcbce983c4f820eb550d6fce05d03bfee