Fix Hermes-Xcode integration
Summary:
Following D46097200, the integration between Hermes and Xcode got broken as Hermes is creating the framework in a different folder.

This is a patch to unblock people internally to build RNTester with Hermes and to run it on a simulator.
We should find a way to parametrize this, passing the target from Xcode to this script, after checking where Hermes is built

## Changelog
[internal] - Fix RNTester-Hermes-Xcode integration for testing locally on simulators

Differential Revision: https://internalfb.com/D46224915

fbshipit-source-id: 4cd684b859875a3c88d98d8e64c7f8ef0b3f753d