W3CPointerEvents: fix NPE due to null targetView (#37785)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37785

Changelog: [Android] [Fixed] - W3CPointerEvents: fix NPE due to null targetView

For reasons not entirely understood, we sometimes end up with a null `targetView` in this code path during the dispatch of an `ACTION_CANCEL`. This results in an NPE during the call to `getChildOffsetRelativeToRoot`. This change adds a null check before the call to avoid the crash.

Reviewed By: cortinico

Differential Revision: D46562263

fbshipit-source-id: a6e8276cfb7213e5a0f64b78bd0ac8b6220e5f0b