Refactor set-rn-template-version
Summary:
Changelog: [Internal]

`set-rn-template-version` used to just set the `react-native` version for the template package (which is nested within `packages/react-native`.

This diff is generalizing this to accept any dictionary of dependencies and their specified version to apply to the template's package.json.

The idea being that we'll pass through that nightly dict eventually. For now, this diff doesn't introduce that behavior, just the ability to and updates it's usages in circleCI and other testing scripts

Reviewed By: cortinico

Differential Revision: D46511859

fbshipit-source-id: 1c9d60561f9811f0546184024cb7a289d64cfa2c