Implement id property in ReadOnlyElement (#37755)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37755

This implements the `id` property in `ReadOnlyElement` as an alias for the `id` or `nativeID` props of the element.

This is currently implemented using an internal API in React but we should find an alternative solution for this in Fabric in the near future. Adding this now because it's useful for debugging.

Changelog: [internal]

Reviewed By: javache

Differential Revision: D46518381

fbshipit-source-id: 37c3c3fcfeab7679a324acb1c05f8177904e06e5