introduce RCTTestUtils (#37810)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37810

Changelog: [Internal]

in this change, i introduce a helper file for swizzling selectors that can be used for building custom shims and fakes. i tried putting this in the cxx target, but i struggled a lot with managing the dependencies and understanding the extra behavior that was happening from being in a `rn_xplat_cxx_library`. this file is also not obj-c++, so i feel like we should decouple if possible.

it does seem like this is a new practice i'm introducing, so let me know if there's something that i'm missing.

Reviewed By: cipolleschi

Differential Revision: D46507803

fbshipit-source-id: 16e8dad7abf30963aa61642b0f2764bcc5204003