Validate number of arguments passed to UIManagerBinding methods (#37783)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37783

We're not checking the number of arguments received before accessing the `arguments` array in JSI methods in `UIManagerBinding`, which could cause access to garbage values. This fixes that by adding a check in all methods to ensure the number of arguments received is at least the arity of the function.

Changelog: [internal]

Reviewed By: rshest

Differential Revision: D46564580

fbshipit-source-id: 7012e0d6fb12c5024e4d6a4c0fb7a8c00263ccac