Add plumbing/boilerplate for an iOS implementation of the `gotpointercapture` and `lostpointercapture` events
Summary:
Changelog: [Internal] - Add plumbing/boilerplate for an iOS implementation of the `gotpointercapture` and `lostpointercapture` events

This diff simply adds the boilerplate necessary to hook up the `gotpointercapture` and `lostpointercapture` events to the fabric iOS touch handler. This diff does not contain any actual implementation of their behavior as that will occur in future diffs.

Differential Revision: https://internalfb.com/D44977499

fbshipit-source-id: 3aa72f9e84b0fce3fc75f0c33a8153479eb1a317