Fix undefined symbols `_FBcoreLocalexxHash48` and `_RCTLocalizedStringFromKey` in OSS (#37838)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37838

We see a linker failure in the OSS build because our header exports C-style functions (not ObjC), that Objective C will assume are C ABI, but we build them as C++ functions.

Internally these are all marked as `extern "C"`, so to not deviate I did the same here to expose these always as plain C functions.

Changelog: [Internal]

Reviewed By: cipolleschi

Differential Revision: D46671506

fbshipit-source-id: f8398b92e82e4e39216c0b45508b19d1df198bba