Add getTestId into views/common (#37851)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37851

## Changelog
[Internal] -

The `testId` value in practice can be used for other things than e2e tests - e.g. there is already a use case for Accessibility, but also we may want to log it as an extra data point with e.g. performance metrics.

This exposes a helper function to extract a `testId` from a view.

Differential Revision: D46684365

fbshipit-source-id: fad17546dded3008d7de332a5610f1045f6ad213