Localize Talkback strings (#37856)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37856

These strings are spoken to the user when using an ascessibility role or state not natively supported by Talkback.

We can enable these for translation by removing flags to disable localization, and adding descriptions to each string. These are derived from MDN for ARIA (used for iOS strings), along with Android and iOS documentation.

Note that this was explicitly disabled as part of D15087192 because of creating spam which was hitting the GitHub repo, but when translation was enabled for the dev menu, this ended up being a non-issue after initial translation. And we really do want these to be localized.

Changelog:
[Android][Fixed] - Localize Talkback strings

Reviewed By: cortinico

Differential Revision: D46692127

fbshipit-source-id: 052791d1e5aa2def6529357860d9db6904351be3