Descriptive CocoaPods error if CMake is not installed (#37839)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37839

The CocoaPods build assumes that a `cmake` command is present during `pod install`. It acts gracefully if it is not and propagates an empty string path, leading to a confusing error during the build, and requiring a `pod deintegrate` after CMake is installed because of when the binary is set.

This replaces the usage with `Pod::Executable::which!`, the built-in CocoaPods method of finding an executable from `PATH` which errors out the build if not found. This will cause a failure during `pod install`, when we are looking for the command, and specify that `cmake` is the missing executable.

This branch is only taken for source builds, so this will not show up in the cases where we do not need CocoaPods.

Changelog:
[iOS][Fixed] - Descriptive Cocoapods Error if CMake is not installed

Reviewed By: cipolleschi

Differential Revision: D46670007

fbshipit-source-id: d508743d51f9b99e1bce7d4d3cef6395d5abf46d