Define FabricUIManager as an interface instead of as an object (#37849)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37849

The typing for `FabricUIManager` makes more sense as an interface (the same way we type native modules) than as an object, as the underlying implementation used to be a host object where properties were actually not enumerable, etc. This replaces that with a definition more aligned with how we actually use it.

This also allows us to extend the mock with testing specific methods more easily (like in D45278720).

Changelog: [internal]

Reviewed By: javache

Differential Revision: D46684044

fbshipit-source-id: 379cc3f95a7cbae733ed6a3a57e68b33efd21d34