Introduce Tuist in the Template
Summary:
[tuist.io](https://tuist.io) is a tool that allows to describe the xcode project-related files using Swift, in a declarative way.

This tool solves the problem of working with the xcodeproj files completely, as it abstracts its details.

The xcodeproj is notoriously a painful file to work with, especially when multiple engineers works together on the same project, and it is also one of the main sources of frustration for our users when it comes to update an app.

This change introduce the `Project.swift` file for the template. However, currently, using tuist has to be done manually, as the RN CLI does not have knowledge of it. Therefore, to make everything work, we are going to commit all the files that tuist generates, to have an atomic commits that works.

## Next Steps:
1. update the CLI so that it can use swift
2. Remove the generated files from the template and add them to the .gitignore.
3. update RN Doctor so that it knows about tuist

## Changelog:
[iOS][Added] - Introduce `tuist` in the template

Differential Revision: D46711973

fbshipit-source-id: dc904d52fba63198439b988a59c749c8ab55d4a8