Add Pointer Capture APIs through intercepting events in UIManagerBinding
Summary:
Changelog: [Internal] - Add Pointer Capture APIs through intercepting events in UIManagerBinding

This implements the basics of pointer capture in a novel way by intercepting events created by the host platform in UIManagerBinding and modifying/emitting additional events sent to JS. Because it lives in the C++ land we can synchronously call the `setPointerCapture`/`releasePointerCapture`/ect. methods from JS without having to risk any deadlocks or thread spinning. Following up on this I'd like to extend this idea further and move more of the functionality we've implemented in ObjC & Java so far to this shared layer to make each of the platform's implementations easier to maintain.

Differential Revision: D46071147

fbshipit-source-id: 61097ffb93103e192be3698296d28bab550568d5