Rollout TurboModuleBinding::Prototype (#38220)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38220

We ran an experiment to test different implementations of TurboModules HostObjects, as the current one has various inefficiencies, such as re-creating HostFunctions on every property access. The strategy we found to be most efficient and flexible longer-term is to represent the TurboModule with a plain JavaScript object and use a HostObject as its prototype. Whenever a property is accessed through the prototype, we cache the property value on the plain object, so it can be efficiently resolved by the VM for future accesses.

Changelog: [General] TurboModules are now exposed to JS as the prototype of a plain JS object, so methods can be cached

Reviewed By: sammy-SC

Differential Revision: D47258286

fbshipit-source-id: 8954e16dafb30551dd8c72cd88f800e6be79cce3