React-yoga podspec (#38345)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38345

Now that Flipper does not depend on Yoga, RN's Yoga pod does not need to conform to the same version, interface, as the standalone Yoga pod.

This changes the `Yoga.podspec` specific to React Native, renaming it to `React-yoga.podspec`. Now that we do not need to worry about a Swift interface, we can expose all of the headers Fabric expects to be able to find as public, for React Native. That effectively undoes the kludges added with D45139075.

Changelog: [Internal]

Differential Revision: D47463111

fbshipit-source-id: a7e258e81c78722cdbf2abe8474d6526422892b6