Interop: Fix regression test (#38476)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38476

The TurboModule interop layer test asserts that SampleLegacyModule.const2 == 390.

But SampleLegacyModule.const2 was actually the screen size of device, which is different locally vs sandcastle.

So, while the test passed locally (b/c screen size = 390), it failed on Sandcastle (b/c screen size = 375).

So, this diff hardcodes const2 to be 390.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: christophpurrer

Differential Revision: D47379531

fbshipit-source-id: 0abf079400d376d4816fecf1e7dac6db6fc688c1