configure turbomodule main queue setup logic (#38544)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38544

Changelog: [Internal]

in this change, i'm setting up the logic where we can disable the behavior that triggers the main queue init of a turbomodule even if it's not explicitly set by the consumer.

the reason i'm looking to do this is because i ultimately want to get rid of this logic. i find the assumptions that this code makes are incorrect, also the behavior is not documented, so it's an opportunity to simplify.

Differential Revision: D47622456

fbshipit-source-id: 3d61e14747393b8e246f0d543904f79d95015bd4