Abstract underlying Color implementation (#38668)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38668

On other platforms (e.g., react-native-windows), it's possible that platform colors may not be efficiently represented as int32_t values. In the case of react-native-windows, Color can either be an ARGB value or a list of fallback strings for platform colors.

This change should decouple how platforms represent color values, allowing them to manage how they are used at the point they are used.

## Changelog:
[General] [Added] - Support customization of underlying Color representation in out-of-tree platforms

Differential Revision: https://internalfb.com/D47873465

fbshipit-source-id: f544cf085153f689dd9fcf3fa8d0e78b985333ce