Allow RCTAppDelegate pod to be imported as a module (#38717)
Summary:
If another pod wants to depend on `React-RCTAppDelegate` and is written in Swift, the user needs to enable modular headers for this pod in the Podfile. This is not very convenient as this cannot be changed as part of the podspec, thus requires additional steps from the user. In Expo, our autolinking just fixes that automatically for all pods that require it (not only for RN's pods).

## Changelog:

[IOS] [CHANGED] - Set DEFINES_MODULE xcconfig in React-RCTAppDelegate to generate a module map for this pod

Pull Request resolved: https://github.com/facebook/react-native/pull/38717

Test Plan:
- rn-tester builds
- `pod install` with a dependency containing Swift code and depending on `React-RCTAppDelegate` no longer requires the user to use modular headers for this pod

Reviewed By: NickGerleman

Differential Revision: D47955835

Pulled By: cipolleschi

fbshipit-source-id: 779516a8396925e52c28b87d6fcf096357333bf5