Do not swapLeftAnd right if not in RTL (#38748)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38748

changelog: [internal]

Android was setting flag `swapLeftAndRightInRTL` to true regardless if the context was RTL or LTR. This causes unnecessary tree traversal + invalidation of all yoga nodes. This is a completely unnecessary work when layout direction is left to right.

To fix this, I made sure Android not longer sets `swapLeftAndRightInRTL` to true and in Fabric we only check the flag for RTL context.

Reviewed By: NickGerleman

Differential Revision: D47913605

fbshipit-source-id: 1d938b0dc9ba16a73b076f626055055162e3495f