Fix compilation warnings related to YogaShadowNode (#38781)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38781

## Changelog:
[Internal] -

These pop up when compiling with `/Wall`, fixes two legitimate warnings around using YogaShadowNode/Props in RN.

The `suggestedIndex` in `ShadowNode::replaceChild` (and the overriden one in `YogaLayoutableShadowNode`) is used in a way around the code that suggests that it **is** expected to be negative (checking for non-negativity, assigning -1 as default etc), so having it as a `size_t` type argument both makes things confusing and generates the warning.

I believe it's a good idea to be consistent and use the same type for the index throughout. In majority of the cases it's `int32_t` as of now.

Reviewed By: rozele

Differential Revision: D48059620

fbshipit-source-id: c4bab91f5c66c62d00947f96cb48906a8499f282