Add MC to disable `onScroll` requiring `scrollEventThrottle` on iOS. (#38742)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38742

If you try to add an `onScroll` event listener on iOS, you will get a warning that you must additionally add a `scrollEventThrottle`, otherwise you will only receive a single event throughout the duration of the gesturing.

Values under 17ms are unthrottled, so it is common to see components passing magic values `16` and `0.0001` to ask for full events.

This behavior is inconsistent with Android, Web, and Windows.

This change runs an experiment on iOS to continuously fire scroll events to users of `onScroll` without `scrollEventThrottle`. Because there are warnings already, I think this case will be rare, and any code written for Android will already have to be written to be compatible with it.

I intentionally left the warning in for now, so that folks don't add new code relying on this QE while it is being tested.

Changelog: [Internal]

Reviewed By: yungsters

Differential Revision: D47975260

fbshipit-source-id: 8f182c5514cc557e1804586e8cbc68fc54502d83