More robust handling of errors when trying to load JS bundle
Summary:
## Changelog:
[Internal] -

When trying to load JS bundle, there were conditions under which it could just silently fail, not giving much to start with when troubleshooting.

It was relying on Folly internals to handle the error, and depending on the context it could either throw an exception or just silently exit the process if exceptions are not enabled in Folly.

This diff makes this error handling more explicit, with a clear error message in the log.

Reviewed By: NickGerleman

Differential Revision: D48147690

fbshipit-source-id: 5a8204a873fd2b68c4510f4662ca6cba8738ab2e