Mitigate flakiness in iOS
Summary:
This diff introduces a script `run_with_retry` that we can apply to single commands in order to mitigate the flakiness.

This can be useful when networking is involved, to retry installing some dependencies, or for example with some e2e/integration tests.

The diff applies this rerun to the iOS tests so we mitigate failures in CI.

## Changelog:
[Internal] - Add script to retry CI steps and mitigate iOS flakyness.

Differential Revision: D48189365

fbshipit-source-id: 944aec018afd23c642efee0ca40ef5aafc254cd7