reuse existing source URL for bridgeless (#38860)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38860

Changelog: [Internal]

so i was thinking that the `getBundleURL` API that we added on the app delegate is redundant since all working apps will have the source url provided in `sourceURLForBridge:`. i briefly perused GH: https://github.com/search?q=sourceURLForBridge&type=code, and it seems no one actually needs the bridge to generate this url.

i think there's an opportunity to remove a step from the upgrade flow by either:
1) getting rid of `getBundleURL` and reuse `sourceURLForBridge:`
2) fallback to `sourceURLForBridge:` in `getBundleURL`

in this diff, i ended up going with 1), but i'm open to options. the tricky thing is that `sourceURLForBridge:` implies the old architecture.

Differential Revision: D48141201

fbshipit-source-id: d7de8012700f05924fa1b024704ca7526a1972cf