feat(rn-tester): Improve the Modal tester
<!-- Thanks for submitting a pull request! We appreciate you spending the time to work on these changes. Please provide enough information so that others can review your pull request. The three fields below are mandatory. -->

 ## Summary:

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

The modal tester in the sandbox was lacking a bit of functionality that
the Modal component exposes -- especially on Android. This change
revamps the Modal page to more-closely resemble the API documentation by
exposing all of the options, and annotating which ones are reserved for
the different platforms.

Additionally, this change puts the modal controls into the created modal
itself. This allows the user to more-easily test what happens if the
props change during the lifespan of the modal (related PR: #38947).

 ## Changelog:

<!-- Help reviewers and the release process by writing your own changelog entry.

Pick one each for the category and type tags:

[ANDROID|GENERAL|IOS|INTERNAL] [BREAKING|ADDED|CHANGED|DEPRECATED|REMOVED|FIXED|SECURITY] - Message

For more details, see:
https://reactnative.dev/contributing/changelogs-in-pull-requests
-->

[INTERNAL] [CHANGED] - Revamp Modal tester in rn-tester

 ## Test Plan:

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes the user interface. -->