Null out ListMetricsAggregator contentLength on orientation change
Summary:
It is a logic error to query for list metrics when a valid list content length has not yet been observed.

An invariant is meant to catch this if it happens, but a non-default orientation causes us to zero out a value which is null, and our null check does not catch this, and we return a nonsense value.

Change from zeroing out the field to nulling it out instead.

Changelog: [Internal]

Differential Revision: D48251645

fbshipit-source-id: 8fad7618f5768bc2f064d47a0a271cf68e38d968