Fix inverted `contentOffset` in scroll events in RTL
Summary:
`UIScrollView` `contentOffset` is flow-relative, so `x` is relative to the right edge of the screen. This coordinate space disagrees with layout events, `scrollTo` coordinates, and other platforms.

This applies the same logic we use for inverting `scrollTo` coordinates to invert `contentOffset` in scroll events, in both Paper and Fabric. We then remove the iOS specific workaround we have in VirtualizedList.

Differential Revision: D48379915

fbshipit-source-id: caa48b99d43298cf51e5d668dd7e7f68fee2ec4e