RN: Change Source Package Versions to 0.0.0
Summary:
Changes the `package.json` for `react-native/monorepo` and `react-native` to 0.0.0 (instead of 1000.0.0).

This reduces the risk of either human or automation error accidentally publishing a v1000.0.0 of the `react-native` package to npm, which would be very problematic.

This should have no visible impact to any existing processes, though.

Changelog:
[Internal]

Differential Revision: D48450601

fbshipit-source-id: 380db364f91c5ede9d80a39562db6a8da1918411