Cleanup BUCK setup for our JVM tests (#39061)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39061

We're suppressing JVM tests using BUCK which is causing misalignment between which tests we run on Buck vs Gradle.
Instead we should be using the Ignore annotation (which we already do) to decide which tests to skip, so that Buck can just pick all the Java/Kotlin file in this folder.

Changelog:
[Internal] [Changed] - Cleanup BUCK setup for our JVM tests

Reviewed By: cipolleschi

Differential Revision: D48433432

fbshipit-source-id: c919d4988f424584d75160c67667352e48cce051