Add ability to get ShadowNodeFamily's ancestors as shared pointers (#38867)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38867

Changelog: [Internal] - Add ability to get ShadowNodeFamily's ancestors as shared pointers

While working on the hover tracking logic I've found the need to recieve a node's ancestor list to contain shared pointers to ShadowNodes instead of references to them. This isn't necessarily to get ownership of the ancestors but instead will allow me to store those shadow node lists as weak pointers so I have a mechanism to know when they've been deallocated.

RFC: This effectively just adds a copy of the existing `getAncestors` method with minor changes. This creates some code duplication but I'm unsure of a more elegant way to do this so I'm open to ideas.

Differential Revision: D48170045

fbshipit-source-id: 158c6349fb39681586f7766318bc9ee0196a829a