Remove usage of HERMES_ENABLED from react-native-xcode.sh (#39128)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39128

`HERMES_ENABLED` was supposed to indicate wether hermes-engine pod is present in the project. It was initialized by grepping `hermes-engine` from `Podfile.lock`.
This is redundant because Hermes is enabled by default and it is sufficient to check that `$USE_HERMES != false` for cases that are Hermes specific.

Changelog: [Internal] - Remove usage of HERMES_ENABLED from react-native-xcode.sh

Differential Revision: D48600174

fbshipit-source-id: 5c0324cb1169e879aba7cd68b987efcdd6c9e5e5