Include accessibilityValue prop values in accessibilityValue (#39184)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39184

Previously, when `props.accessibilityValue` was enabled and other props were enabled, like `props.accessibilityRole`, that are used in the accessibilityValue for iOS, the value from `props.accessibilityValue` wasn't announced.

This is a problem for cases like radio button, which also needs to announce the placement of the radio button within a group, e.g. "[label], radio button, checked, 1 out of 3".

This diff fixes this problem by including the value from `props.acessibilityValue` within the method that builds the `accessibiityValue` for VoiceOver.

Reviewed By: cipolleschi

Differential Revision: D48534155

fbshipit-source-id: 176ba76ff772741a5cdc7d2a56c6cdfb92674477