Fork ReactCommon/hermes/inspector as inspector-modern (#39253)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39253

X-link: https://github.com/facebook/hermes/pull/1112

Changelog: [Internal]

Forks `ReactCommon/hermes/inspector` → `ReactCommon/hermes/inspector-modern`. More changes to `inspector-modern` will come in subsequent diffs.

The Gradle, CocoaPods and Buck builds of React Native have been updated to use `inspector-modern` everywhere. The code in `ReactCommon/hermes/inspector` is thus **deprecated** and unused in the open source build, and will be deleted in an upcoming diff (likely before the 0.73 cut).

NOTE: The reason we're not immediately deleting `ReactCommon/hermes/inspector` is that there are a handful of references to it in Meta's internal monorepo (*outside of React Native*) that we are in the process of migrating away. Once we've deleted the existing `ReactCommon/hermes/inspector`, we *may* rename `inspector-modern` back to `inspector`.

Differential Revision: D48897203

fbshipit-source-id: dd3c262c9ca0bb88163bfcf5cc51826085fed5e1