Only disable throttling when scrollEventThrottle isn't set (#39293)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39293

This QE is meant to change behavior to not throttle if `scrollEventThrottle` isn't set, but it's actually disabling throttling entirely.

This changes the gating to instead change the initialization path to set `_scrollEventThrottle = 0` when the QE is set.

`_scrollEventThrottle` is already set to zero when the `scrollEventThrottle` prop is removed/set to null, as the default value in ScrollViewProps.

Changelog: [internal]

Reviewed By: sammy-SC

Differential Revision: D48968754

fbshipit-source-id: c46c7f5093a60e326267c2e5f2f86dc2d545ac7f