Fork jsinspector as jsinspector-modern (#39288)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39288

Forks `ReactCommon/jsinspector` → `ReactCommon/jsinspector-modern`.

The Gradle, CocoaPods and Buck builds of React Native have been updated to use `jsinspector` everywhere (outside of `ReactCommon/hermes/inspector`, which is itself dead code as of D48897203). The code in `ReactCommon/jsinspector` is thus deprecated and unused in the open source build, and will be deleted in an upcoming diff (likely before the 0.73 cut).

Changelog: [Internal]

Reviewed By: blakef

Differential Revision: D48966244

fbshipit-source-id: db81739da83ef9be73690d95d010065e3b3441c7