Implement clientTop/clientLeft in ReadOnlyElement (#39308)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39308

This adds a new method in Fabric to get the border size for an element, and uses it to implement the following methods as defined in https://github.com/react-native-community/discussions-and-proposals/pull/607 :
* `clientLeft`: left border width of the element.
* `clientTop`: top border width of the element.
If the element isn't displayed or it has display: inline, it return 0 in both cases.

These APIs provide rounded integers.

Changelog: [internal]

Reviewed By: mdvacca

Differential Revision: D49009140

fbshipit-source-id: 7cc2fac1ec0526a5ad441bf71039333e10ff9696