Use Hermes-engine version in caches for RNTester jobs (#39363)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39363

With the recent CI optimization, we were caching the hermes-engine stored in the Pods folder in order to reuse it.
However, by doing so, we would not avtually use the hermes-engine we were building in CI nor we were using the most recent version in the E2E tests as the Podfile.lock would not have actually changed and Cocoapods would have found a proper version for the Hermes-engine.

## Changelog:
[Internal] - use the hermes engine version to invalidate cocoapods caches

Differential Revision: https://internalfb.com/D49125000

fbshipit-source-id: 57e427886ae7625cd8d7a6a056e61b5e21ad4c2e