Breaking: Remove YGConfigGetInstanceCount
Summary:
This was added in https://github.com/facebook/yoga/pull/497 specifically for tests related to memory leaks in the C# bindings to count how often YGConfigFree.

This is the wrong layer for this check, we don't have officially supported C# bindings anymore, and this API is not safe when Yoga runs on multiple threads. This removes it.

Differential Revision: D49131207

fbshipit-source-id: 7c51fcb3aefea0a4d8217c474477a664b7bfd78f