Fixup exports
Summary:
D18029030 added `-fvisibility-hidden`, and a corresponding `YOGA_EXPORT` macro for defining shared library visibility.

There was already a `WIN_EXPORT` macro doing the same thing when building a DLL, but in the header, instead of CPP files, and sometimes missing for important parts of the public API?

This reconciles those into a single visiblity macro, that we always place with declaration instead of definition.

Differential Revision: D49132643

fbshipit-source-id: 3c6a5785aa4a202985676f9d0c06baae9bdf1ca3