Deprecate default_flags in Podfile (#39389)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39389

The Podfile of the Template is using a function called `get_default_flags` to get the default flags.
Its behavior is duplicated in both the default value of the `use_react_native!` function and in the body of the same function, making that helper actually redundant.

In this change, we are deprecating it so we can remove it in 0.74 with no breakages.

## Changelog:
[iOS][Deprecated] - Deprecate `get_default_flags` in Ruby scripts

Reviewed By: dmytrorykun

Differential Revision: D49147290

fbshipit-source-id: c97279008c4690882d1edbbce396ce8d67629452