Deprecate default_flags in Podfile
Summary:
The Podfile of the Template is using a function called `get_default_flags` to get the default flags.
Its behavior is duplicated in both the default value of the `use_react_native!` function and in the body of the same function, making that helper actually redundant.

In this change, we are deprecating it so we can remove it in 0.74 with no breakages.

Differential Revision: D49147290

fbshipit-source-id: 858134e777671b104088491d572161f9b7dac6ff