Remove config threading
Summary:
We thread a config through the root node, to every function, that we don't actually use (with the exception of `canUseCachedMeasurement`, and shouldn't, with a model of per-node config.

`canUseCachedMeasurement` does rely on pointScaleFactor, which Fabric previously cleverly decided it would only set on the root nodes config, but that should be fixed now, and I will be making a followup to make `pointScaleFactor` properly act per-node.

Differential Revision: D49180385

fbshipit-source-id: 27c8c20a8ba1b96b48193021f03ed96a51e4c225