Use fbsource clang-format config (#39434)
Summary:
X-link: https://github.com/facebook/yoga/pull/1381

Pull Request resolved: https://github.com/facebook/react-native/pull/39434

This mirrors the clang-format config used by fbsource to Yoga.

They are pretty similar, except for an annoying habit where Yoga's previous forced small functions in headers to be a a single line, so you would get a combination of multiline and single line functions next to each other which are hard to read. That is what motivated this change.

It also enforces header ordering (yay). I don't think we have any side-effect causing headers, so this should be safe.

Reviewed By: yungsters

Differential Revision: D49248994

fbshipit-source-id: 99f4ba5de99e93ae737af6ab08c0ba73dc8f7b64