Annotate HybridData with DoNotStrip (#39342)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39342

Annotate HybridData with DoNotStrip

The stack of diffs aims to introduce new features and deprecate old ones in the React Native Android architecture. The main goal is to make the architecture more stable and easier to use for developers. The changes include marking ReactPackage as NullSafe, introducing the new getModule method, introducing the WillBeDeprecatedInNewArchitecture annotation, deprecating createNativeModules and marking getModule as UnstableReactNativeAPI, and fixing a lint warning in MainReactPackage.

# This diff
This diff is adding the annotation `DoNotStrip` to the `HybridData` field in the `NativeMethodCallInvokerHolderImpl` and `CallInvokerHolderImpl` classes. This is to prevent the Android ProGuard optimizer from stripping the `HybridData` field, which is necessary for the proper functioning of the TurboModule system in React Native. Additionally, the `BUCK` file for the `core` module is updated to include the `androidx:annotation` and `com.facebook.proguard.annotations:annotations` dependencies.

changelog: [intenral] internal# Context

Reviewed By: cortinico

Differential Revision: D49077065

fbshipit-source-id: cb36c0a11bdb8f5f0f4eaa7db1f0c9dcb3be2341