Update useAnimatedProps.js
This commit updates useAnimatedPropsLifecycle to first attach the animated node instantly, rather than waiting on useLayoutEffect. This is done only once (by adding hasMountedRef) and also adds another ref (hasInitiallyAttached) which is checked in useLayoutEffect, to avoid running node.__attach() twice before the first render.

This seems technically equivalent to the UNSAFE_componentWillMount functionality that was in createAnimatedComponent.js prior to RN 0.72 and the move to the updated useAnimatedProps.js.

Fixes an issue (#38510) where changing the key of an animated component would result in a reset of animated values (and basically not re-attaching in time for *native* animations)