Enable Fabric interop in bridgeless mode on iOS (#39495)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39495

We've put `[self unstable_registerLegacyComponents]` inside `#if RCT_NEW_ARCH_ENABLED` block, but we've missed the fact that it was inside the `else` clause of the `if (enableBridgeless)` statement. So it wasn't actually enabled for bridgeless mode regardless of `RCT_NEW_ARCH_ENABLED`.
This diff fixes that, and registers legacy components for the bridgeless mode as well.

Changelog: [Internal]

Reviewed By: NickGerleman, cipolleschi

Differential Revision: D49323913

fbshipit-source-id: 5b8e4ffc6ea1cdf5bbdecbddd5274d70c9b2eeab