Add Systrace instrumentation to RuntimeScheduler (#37880)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37880

This adds systrace sections for the most relevant parts of `RuntimeScheduler`. This helps us identify how things are scheduled, which in this case makes it obvious we're not dispatching events the most efficient way (as top-level callbacks in the runtime executor instead of as tasks in the scheduler).

Changelog: [internal]

Reviewed By: rshest

Differential Revision: D46556399

fbshipit-source-id: aa00d9b1170c8b51abb51462596b57213ec02b4e