Bump template cache to fix iOS CI (#39560)
Summary:
Earlier today we bumped boost and removed a fix for Xcode 15.
Boost is a pod we don't consume from source and for the template we don't have a Podfile.lock to update.
However, the pod we don't control are cached as they don't change often.

This PR fixes bumps the cache to fix the CI issue

bypass-github-export-checks

## Changelog:
[Internal] - Bump iOS CI cache keys for the template

Pull Request resolved: https://github.com/facebook/react-native/pull/39560

Test Plan: CircleCI is green

Reviewed By: christophpurrer

Differential Revision: D49458677

Pulled By: cipolleschi

fbshipit-source-id: 957bb5bcc5fe940b9ea08c3f0b9031a39cee37ff