Static view configs take precedence over native view configs (#39534)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39534

Fetching native view configs is computationally intensive work, so if some component has both native and static VCs, we should try static first.
This diff changes the `NativeComponentRegistry` so it tries to get static view config first, and then native view config as a fallback.
This diff also removes setting `RuntimeConfigProvider` from RNTester. As it was there for the same purpose of inverting that precedence.
Changelog: [Internal]

Reviewed By: cortinico

Differential Revision: D49413562

fbshipit-source-id: d65776cd43eaa97d880b76385423c65e4e8744ba