Make template pods resilient to changes to internal podspec (#39601)
Summary:
With this small change, we are binding the cache for the template to the cache of RNTester.
This is helpful because, by default, templates don't have a Podfile.lock of their own. It is normal, as they are app created from the template and we don't enforce the initial
lock of those dependencies.

However, this can cause some hassle when we bump some internal podspec as the cache that is restored in CI will be outdated.
With this change, the podfile.lock for RNTester is different --  we push it when we change the internal podspecs -- so the cache for the template will be invalidated too.

## Changelog:
[Internal] - Invalidate Template caches when we bump internal podspecs

Pull Request resolved: https://github.com/facebook/react-native/pull/39601

Test Plan: CircleCI is green

Reviewed By: christophpurrer

Differential Revision: D49536315

Pulled By: cipolleschi

fbshipit-source-id: 04f1952f029ea643f2d82d2d690fad147faace05