Reduce visibility of JSIModuleHolder (#39741)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39741

In an attempt to reduce footprint of React Native Android public APIs we are reducing visibility of classes and interfaces that are not meant to be used publicly OR are public but have no usages.
As part of our analysis, which involved looking for usages inside the Meta codebase and code search in OSS, we've detected that this class/interface is public but it's not used from other packages.

If you are using this class or interface please comment in this PR and we will restate the public access.

changelog: [Android][Changed] Reducing visibility of JSIModuleHolder

Reviewed By: RSNara

Differential Revision: D49752136

fbshipit-source-id: 2314d4cc5e58706240efe4502d8f24cce7d182a9