Reduce visibility of PerformanceCounter (#39746)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39746

In an attempt to reduce footprint of React Native Android public APIs we are reducing visibility of classes and interfaces that are not meant to be used publicly OR are public but have no usages.
As part of our analysis, which involved looking for usages inside the Meta codebase and code search in OSS, we've detected that this class/interface is public but it's not used from other packages.

If you are using this class or interface please comment in this PR and we will restate the public access.

changelog: [Android][Changed] Reducing visibility of PerformanceCounter

Reviewed By: arushikesarwani94

Differential Revision: D49803283

fbshipit-source-id: 69f0d4049e8180286963de73c9c891453c81c40b