Try using modern Android libc++ in Yoga and React Native
Summary:
Android NDK 25 uses a version of libc++ that is more than three years old, missing a lot of basic features of C++ 20. This is rectified in NDK 26 (latest LTS NDK), which brings us up to date with latest Clang, and later NDK versions will now also bump libc++ as part of bumping LLVM/Clang.

This requires an Alpha AGP version, which is... spooky, and we would need to update that before shipping RN 0.74/Yoga 3.0. It does bring us closer to what we want to ship then however.

Differential Revision: https://internalfb.com/D49895949

fbshipit-source-id: 7e78cd00455a10af9c8b1473c9a0a309f7f589cb