Expose Qpl.isMarkerOn to React Native (#39830)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39830

## Changelog:
[Internal] -

Add `QuickPerformanceLogger.isMarkerOn` API, which allows to check whether the given QPL marker is going to be sent to the server or not (the latter may happen due to e.g. downsampling).

This allows to avoid some extra unneeded overhead when logging QPL events in some heavily sampled scenarios.

Reviewed By: rubennorte

Differential Revision: D49949527

fbshipit-source-id: 9d7f93beee45d498c799a94b16cd7c68ec1a9340