update documentation for UIManagerProvider interface (#39762)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39762

In an attempt to reduce footprint of React Native Android public APIs we are reducing visibility of classes and interfaces that are not meant to be used publicly OR are public but have no usages.
As part of our analysis, which involved looking for usages inside the Meta codebase and code search in OSS, we've detected that this class/interface is public but it's not used from other packages.

If you are using this class or interface please comment in this PR and we will restate the public access.

changelog: [Android][Changed] Reducing visibility of UIManagerProvider

bypass-github-export-checks

Reviewed By: cortinico

Differential Revision: D49803268

fbshipit-source-id: 8bc261115beae8fe0666ed95b65b01c1772a5d15