Reduce visibility of CancellationTokenRegistration
Summary:
In an attempt to reduce footprint of React Native Android public APIs we are reducing visibility of classes and interfaces that are not meant to be used publicly OR are public but have no usages.
As part of our analysis, which involved looking for usages inside the Meta codebase and code search in OSS, we've detected that this class/interface is public but it's not used from other packages.

If you are using this class or interface please comment in this PR and we will restate the public access.

bypass-github-export-checks

changelog: [Android][Changed] Reducing visibility of CancellationTokenRegistration

Reviewed By: arushikesarwani94

Differential Revision: D49803300

fbshipit-source-id: c7992bdaa0994c1035894a15af1e9d0bf7ebd204