Make PerformanceObserver report correctly supported performance entry types (#39951)
Summary:

## Changelog:
[Internal] -

This makes PerformanceObserver API more robust in regards of telling which exactly types of the performance entries are supported.

At this point, we either tell that we support mark/measure/event ones on the New Architecture, or none otherwise. The source of truth of this information should be on the native side.

Reviewed By: rubennorte

Differential Revision: D50010982