Fix react.dev docs urls (#40095)
Summary:
All "https://react.dev/link" links give a 404 and are deprecated. I replaced all URLs with new ones or, if this was not possible, replaced them with a link to legacy documentation.

## Changelog:

[INTERNAL] [FIXED] - Fixed links to React documentation

Pull Request resolved: https://github.com/facebook/react-native/pull/40095

Test Plan: All of the old links didn't work. I tried to match the content of the errors to the new documentation as best as possible. Current links have been tested and direct you to the most relevant article section.

Reviewed By: christophpurrer

Differential Revision: D50094451

Pulled By: arushikesarwani94

fbshipit-source-id: 79fd9e729495cadfb067d94fb58acb30ca308347