Deterministic onLayout event ordering for iOS Paper
Summary:
The ordering of `onLayout` events is non-deterministic on iOS, due to nodes being added to an `NSHashTable` before iteration, instead of an ordered collection.

We don't do any lookups on the collection, so I think this was chosen over `NSMutableArray` for the sake of `[NSHashTable weakObjectsHashTable]`, to avoid retain/release. Using a collection which does retain/release seems to cause a crash due to double release or similar, so those semantics seem intentional (though I'm not super familiar with the model here).

We can replicate the memory semantics with ordering by using `NSPointerArray`. This change does that, so we get consistetly top-down layout events (matching Fabric, and Android Paper after a recent change). This lets us use multiple layout events to calculate right/bottom edge insets deterministically.

Changelog:
[iOS][Changed] -  Deterministic onLayout event ordering for iOS Paper

Differential Revision: D50093411

fbshipit-source-id: d9229f98d8032673707fb8b9e0fd169f660edb08