Picking: 'Fix incorrect values for modal sizing'
Originally from https://github.com/discord/react-native/pull/45

This fixes an issues where the fabric renderer would get the correct
sizing via onSizeChanged, write it to the shadow tree generating a state
update callback and then setting the wrong size using ModalHostManager.
Once onSizeChanged is fired by the OS it is treated as the definitive
source of truth for sizing.