Add errata for fix to marginStart/End for row-reverse flex direction (#40803)
Summary:
X-link: https://github.com/facebook/yoga/pull/1419

X-link: https://github.com/facebook/litho/pull/961


This stack is ultimately aiming to solve https://github.com/facebook/yoga/issues/1208

This adds an value to the Errata enum. I will use this to gate this fix as there is potential for users to rely on this bug or have a hack in place to fix it and this would be a breaking change.

Reviewed By: NickGerleman

Differential Revision: D50145273