Fix event type mismatch in FlatList (#41006)
Summary:

changelog: [internal]

Props passed to `_defaultRenderScrollComponent` are props for <ScrollView />. <ScrollView /> has event `onContentSizeChange` but <View /> does not. But `_defaultRenderScrollComponent` returns <View /> if <FlatList /> is nested within the same orientation. `onContentSizeChange` is important event, it is used to determine how many cells should be shown.

How come it used to work?
This is very context dependent. Any update after initial render will fix it. There are many things that need to align for the bug to manifest.

Reviewed By: javache

Differential Revision: D50320323