remove unused methodQueue from RCTImageStoreManager
Summary:
Changelog: [iOS][Breaking] You cannot call methodQueue on RCTHTTPRequestHandler

the synthesize methodQueue API is confusing, it looks like an API only for use within native module implementation, but it's actually needed to create a selector that corresponds to the property declared in the RCTBridgeModule public protocol.
no one is using the methodQueue  on RCTHTTPRequestHandler, so let's get rid of the public access to it.

Differential Revision: D50525896