Pass reactNativeConfig to HermesInstance when not using the native module version (#41082)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41082

We're testing a method to access `ReactNativeConfig` without a dependency on native modules, so we can access it before that infra is initialized in places like Hermes or RuntimeScheduler.

When we're in that variant, this passes the configuration to Hermes so we can use it to set flags in the runtime (like enabling microtasks in D50177355).

Changelog: [internal]

Reviewed By: sammy-SC

Differential Revision: D50450488

fbshipit-source-id: 77f0369f93bb7175c569d51b0569669552a13acf