remove unused methodQueue from RCTHTTPRequestHandler (#41136)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41136

Changelog: [iOS][Breaking] You cannot call methodQueue on RCTHTTPRequestHandler

the `synthesize methodQueue` API is confusing, it looks like an API only for use within native module implementation, but it's actually needed to create a selector that corresponds to the property declared in the `RCTBridgeModule` public protocol.

no one is using the `methodQueue`  on `RCTHTTPRequestHandler`, so let's get rid of the public access to it.

Reviewed By: javache, cipolleschi

Differential Revision: D50525900

fbshipit-source-id: f83738491d0eadc71a6dc3194ee16fe7c8748263