Fix autoscrollToTopThreshold on iOS old arch (#38245)
Summary:
Currently `autoscrollToTopThreshold` does not work correctly, it will scroll to top even when the scroll position is past the `autoscrollToTopThreshold` value.

The value of x/y is taken before we adjust the scroll position so we do not need to subtract the delta.

Note that this was already fixed when I ported this code to fabric, so this fix is only needed in the old arch code.

bypass-github-export-checks

## Changelog:

[IOS] [FIXED] - Fix autoscrollToTopThreshold on iOS old arch

Pull Request resolved: https://github.com/facebook/react-native/pull/38245

Test Plan:
In RNTester example, threshold is set to 10, so it should not scroll to top if we are further than 10px from the top of the list.

Before:

https://github.com/facebook/react-native/assets/2677334/13723787-1bc4-4263-9bcb-91ddf7454de3

After:

https://github.com/facebook/react-native/assets/2677334/a8cfdaac-59fc-40de-970a-ff992366e25f

Reviewed By: rshest

Differential Revision: D50447644

Pulled By: cipolleschi

fbshipit-source-id: b21f1836db293120a7a795c8f8f6dd54887495a7