Refactor LazyTurboModuleManagerDelegate to avoid holding references to TurboModules (#41181)
Summary:

The goal of this refactor is to ensure that LazyTurboModuleManagerDelegate doesn't hold references to TurboModules.

EveryTime that LazyTurboModuleManagerDelegate.getModule is called, it will create a new TurboModule. This 'should' be fine because the references to already created TurboModules are held on
TurboModuleManager.mModuleHolders.

As part of this diff I'm also throwing an exception when the method LazyTurboModuleManagerDelegate.unstable_isModuleRegistered is called. This should be fine because I ensured that
"LazyTurboModuleManagerDelegate.unstable_isModuleRegistered' is not called for this experiment.

changelog: [internal] internal

Reviewed By: RSNara

Differential Revision: D50610163