Add add_dependency function to ract_native_pods and Migrate React-RCTAppDelegate (#41355)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41355

This change expose a new API to react_native_pod to add dependencies that automatically configure their search paths when using frameworksa and with multiple Apple platforms.
It also migrates React-RCTAppDelegate to this new mechanism to test that it works.

## Context
Last week I helped macOS to work with static framework.
When multiple platforms are specified, frameworks are build in two variants, the iOS and macOS one.

This break all the HEADER_SEARCH_PATHS as now we have to properly specify the base folder from which the search path is generated.
See also [this PR](https://github.com/microsoft/react-native-macos/pull/1967) where I manually make MacOS work with `use_framewroks!`

## Changelog:
[Internal] - Add helper function to create header_search_path

Differential Revision: https://internalfb.com/D51029484

fbshipit-source-id: ff39bb51a0b76c5139dc7f174ea49df7b2d54df4