Use flexStart functions instead of inlineStart ones in cases where setting position (#41413)
Summary:
X-link: https://github.com/facebook/yoga/pull/1464


In all cases where we had to set a position we define what edge to set it on. All calls here set the flex edge, not the inline one. That makes sense because the flex start edge is most important when applying some of the key parts of the flexbox algorithm, which is in charge of deciding the position for the most part (we resolve flex direction to RowReverse/Row when in RTL). While all the inline stuff is mainly used to set the sides for specific layout values like margin, border, and padding.

I also added a test case here for 2 of the callsites changed, but the rest either fail for different reasons or get overriden due to the remaining logic. Still I think we should be consistent.

Reviewed By: NickGerleman

Differential Revision: D51187444