Weak SurfaceHandler Link to ShadowTree (#41552)
Summary:

`SurfaceHandler` keeps a raw pointer to ShadowTree and UIManager. This is normally set to null when a surface is stopped, but it seems like there may be a race condition where the `ShadowTree` dies during surface start, and we call something which assumes the pointer is valid.

This converts the raw pointer to weak pointer, so we get nulled out automatically, and inserts a null-check in the case where we seem to be accessing invalid ShadowTree.

Changelog: [Internal]

Differential Revision: D51452801