Terminate on logic error in createNode HostFunction (#41567)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41567

Sometimes `createNode` calls into is throwing `std::out_of_range_error` or `std::length_error` in response to both vector and string operations.

Instead of propagating `std::logic_error`, which indicates a defect in native code, terminate, so we can get an actionable native stack trace.

`createNode` and `cloneNode` also both ocasionally see `bad_alloc`, but this is not usually an instance of a defect at the allocation-site, and throwing would be more graceful.

Changelog:
[Internal]

Reviewed By: javache

Differential Revision: D51463600

fbshipit-source-id: 870cbf3538d8ccbc01ded2868781a63ba12a941c