Don't enable `ordinalCount()` or `bitCount()` for bitflag enums
Summary:
Bitfield enums are not sequential, so use of these functions on these enums would be invalid.

I looked at whether we could trivially move `bitCount` to template based on `ordinalCount`. `bitCount` must be constexpr, since we use it directly as a bit-field size constant. `log2` and `ceil` to be constexpr, which isn't in stock C++ until C++ 26 🥲.

Differential Revision: D51518899