refactor[AppContainer]: remove innerView ref from state
Summary:
We were keeping ref in the AppContainer's state before migrating these components to functions.

We actually need a real ref here, so we can use `useImperativeHandle` on it later. These refs will be passed as an argument for DebuggingRegistry subscription, whichi will call them on all necessary AppContainers to highlight required components.

Differential Revision: D51536772