avoid scheduling frame callback if there are no events (#41658)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41658

changelog: [internal]

FabricEventDispatcher does not need to run on every frame. Whenever a new event is added to Fabric's event queue, it will call `FabricUIManager.onRequestEventBeat` in Java. This in turn calls `FabricEventDispatcher.maybePostFrameCallbackFromNonUI` and adds a frame callback on the Choreographer.

This makes code simpler, as we do not need to manage the frame callback subscription.

Reviewed By: sammy-SC

Differential Revision: D50604303

fbshipit-source-id: ce2c7b77678bfc14aa7ecac71e40f78263c7036a