Fix issue where percentages were off of the border box, not padding box (#41686)
Summary:
X-link: https://github.com/facebook/yoga/pull/1485


The size of the containing block is the size of the padding box of the containing node for absolute nodes. We were looking at  `containingNode->getLayout().measuredDimension(Dimension::Width)` which is the border box. So we need to subtract the border from this.

Added a test that was failing before this change as well

Reviewed By: NickGerleman

Differential Revision: D51330526