Fix issue where we were not applying flex end correctly when justifying (#41691)
Summary:
X-link: https://github.com/facebook/yoga/pull/1487


The code here was just wrong. I changed it to be the same logic as the Justify:FlexStart case, but with the flex end sides. Then I get the position for the opposite edge since we need to write to flex start side.

Reviewed By: NickGerleman

Differential Revision: D51383792