Convert ReactAndroid and RN-Tester to Kotlin DSL (#41834)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41834

I'm updating other two `build.gradle` to `build.gradle.kts` files.
The only functional change I made was to remove the function to check if major >= 1 and turn on New Architecture. This needs to be moved to RNGP as Kotlin doesn't have dynamic accessors to Object so we can't convert that function.

Changelog:
[Internal] [Changed] - Convert ReactAndroid and RN-Tester to Kotlin DSL

Reviewed By: mdvacca

Differential Revision: D51856356

fbshipit-source-id: ef75579cd3ec121ef6ac9a357c1e10bcf9995432