add back ComponentDescriptor::adopt with shared_ptr (#41670)
Summary:

changelog: [Deprecated] Deprecate ComponentDescriptor::adopt with shared_ptr and use new overload passing in a reference.

In D49054000 I changed the signature of `ComponentDescriptor::adopt`, which is a breaking change.
Here, I'm adding it back and marking it as deprecated with option for people to migrate to the new overload of `ComponentDescriptor::adopt`.

Differential Revision: D51618172