Change offsetParent, offsetTop and offsetLeft to respect positioned ancestor, not parent (#41903)
Summary:

These offset methods are supposed to be in reference to the node's nearest positioned (non-static) ancestor: https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/offsetParent. Right now, because static did not exist, they return the offset from the parent. This changes it so that the API is spec compliant and will look at the position of its ancestors before settling on the right offset. I added a helper `getNewestPositionedAncestorOfShadowNode` to get the correct node. Then I use that to calculate the proper offset

Changelog: [Internal]

Reviewed By: rubennorte, NickGerleman

Differential Revision: D51414950