Introduce "codegenConfig.isLibrary" property (#41655)
Summary:

This diff adds support for checked-in codegen artifacts for libraries.
It introduces a new property to `coegenConfig`, called `isLibrary`. If codegen sees `isLibrary: true` in a project's dependency, it assumes that the library has codegen artifacts in it, and will not generate any code.

Changelog: [General][Added] - Introduce "codegenConfig.isLibrary" property.

Reviewed By: cipolleschi

Differential Revision: D51207265