fix(rn-tester): add missing params when running Android app from watch mode (#41979)
Summary:
Passed `--mode HermesDebug` to `run-android` command when running from watch mode by pressing `a` on terminal running a dev server. The flag is the same as in the `package.json`:
https://github.com/facebook/react-native/blob/27f38f6f0647ec1809ee0a0d8e9da3a77a9115b1/packages/rn-tester/package.json#L17

## Changelog:

[INTERNAL] [CHANGED] - Add missing params when running Android app from watch mode by pressing `a`

Pull Request resolved: https://github.com/facebook/react-native/pull/41979

Test Plan: Run `yarn start` press `a` in the watch mode, and Android app should be build and launch correctly.

Reviewed By: cipolleschi

Differential Revision: D52265462

Pulled By: lunaleaps

fbshipit-source-id: b2fbe6c889d8067876e160a8ce64dedcc4ce24d7