Remove default yoga style extranious code (#42190)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42190

Essentially undoing D51182861 that was put in place to support adding a default position type to Fabric. Previously I had already removed the code that set the default to something other than Yoga default, but I did not remove all this extra code that allows you to do that. Since we no longer need this we should remove it so as not to encourage messing with the defaults in such a way that they differ from Yoga defaults.

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D52515993

fbshipit-source-id: fb4ab726cf73bf08fd6aa44b99196962a9839694