BridgelessUIManager: Finish setLayoutAnimationEnabledExperimental
Summary:
In open source, with the new architecture, layout animations are **always** enabled.

They cannot be disabled.

Therefore, when this UIManagerModule method is called with false, just report an error. That way, if layout animations were explicitly disabled on Android, the developer will know, when they try to enable the new architecture.

Changelog: [Internal]

Differential Revision: D52349297

fbshipit-source-id: 29660746115c36046905116a546a8ecbaacccb94