feat[DebuggingOverlay]: support native commands for highlighting elements (#41745)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41745

Changelog: [Internal]

Support `highlightElements` and `clearElementsHighlights` commands in `DebuggingOverlay` native components.

These later will be used for highlighting inspected component in React DevTools. These commands unblock highlighting elements on the native side, currently we do it on JS side and it mutates the React tree.

We still need to serialize the array before passing it to the native command, because codegen doesn't support it yet.

Reviewed By: javache

Differential Revision: D51603861

fbshipit-source-id: 4f851e17e77ecc66251d622cd96fa4d85d6cb36f