[skip ci] Remove static-behaves-like-relative errata (#42315)
Summary:
X-link: https://github.com/facebook/yoga/pull/1556


Since we aim to ship static to all users of yoga (not just XPR), we need to remove the errata that is gating most of the features. This should be a non breaking change. To ensure that, I added a new errata which, if on, will use the inner size of the containing node as the containing block. This is how it has been for a while and resolving this is risky and time consuming so for the time being we will stick with that.

Differential Revision: D52706161