[skip ci] Add 'static' as position option in style sheets (#42316)
Summary:

We need to change the typing to allow for 'static'. An issue here is that Paper will not have `static` due to missing z-index logic. Unfortunately, we cannot create a fabric-only version of the typing as we cannot have conditional elements of the same name in ts. To remedy this we took out the parsing of the string 'static' in Paper. Instead we will just emit a warning and default to `relative`. 

bypass-github-export-checks

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D51431524