Remove Content-Length header from proxy inspector response
This change removes Content-Length header from proxy inspector response.

This header was resulting in the response being cropped some times, because of erroneously calculated langth. The Content-Length value represents the number of bytes in the response while string.length was used to calculate the value, which gives the number of characters in a string instead of its size in bytes. Specifically, if there are some UTF characters in the string that occupy more than byte, there would be a mismatch in this size which resulted in the response being cropped. This could happen for example when the simulator name contains such character.