Don't inject synthetic Debugger.disable message in modern CDP backend (#42400)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42400

Changelog: [Internal]

TSIA - iOS counterpart of D52040149 on Android.

The overarching principle is that nothing outside of an Agent should be doing anything with the CDP message stream. Here we have a case of `RCTInspectorDevServerHelper` basically impersonating the CDP frontend in order to paper over an apparent lifetime management bug in the old backend; this gets in the way of implementing reloads natively so we disable it under the new backend.

NOTE: I'm gating out both the call site in `RCTBridge` (to signal intent) *and* the actual body of `disableDebugger` (in case any out-of-tree code happens to be using this method).

Reviewed By: voideanvalue

Differential Revision: D50967799

fbshipit-source-id: 759718bf155b8b16c7db54ac2d2507bc71c93436