Add standards based lexer/tokenizer for CSS values
Summary:
Right now this is done by some combination of string splitting, JS side regexing, etc. This will not scale for math expressions, and gets more and more annoying when we try to add more.

In preparation for adding more units, this adds a tokenizer/lexer for a subset of CSS grammar, based on the spec.

This is not hooked up to anything yet, and doesn't add a parser.

Changelog: [Internal]

Differential Revision: D53030661